chore: rename verify export
